# marketplace-api/src/main/resources/openapi/marketplace.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: Marketplace homework API with contract-first approach.
servers:
  - url: http://localhost:8080

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '401':
          $ref: '#/components/responses/TokenInvalidError'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '401':
          $ref: '#/components/responses/RefreshTokenInvalidError'

  /products:
    post:
      tags: [Products]
      summary: Create product (SELLER own, ADMIN any)
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'

    get:
      tags: [Products]
      summary: List products with filters and pagination
      operationId: listProducts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalidError'

  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by id
      operationId: getProductById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/ProductNotFoundError'
        '401':
          $ref: '#/components/responses/TokenInvalidError'

    put:
      tags: [Products]
      summary: Update product (SELLER own, ADMIN any)
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/ProductNotFoundError'

    delete:
      tags: [Products]
      summary: Archive product (soft-delete)
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
      responses:
        '204':
          description: Archived
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/ProductNotFoundError'

  /orders:
    post:
      tags: [Orders]
      summary: Create order (USER/ADMIN)
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'
        '404':
          $ref: '#/components/responses/ProductNotFoundError'
        '409':
          description: Business conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Promo code errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Order rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by id (owner or ADMIN)
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          description: Access denied or ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/OrderNotFoundError'

    put:
      tags: [Orders]
      summary: Update order items (owner in CREATED state or ADMIN)
      operationId: updateOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          description: Access denied or ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/OrderNotFoundError'
        '409':
          description: Invalid state / stock / product inactive / active order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Promo code errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Order update rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel order (owner in CREATED or PAYMENT_PENDING, or ADMIN)
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: Canceled order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          description: Access denied or ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/OrderNotFoundError'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /promo-codes:
    post:
      tags: [PromoCodes]
      summary: Create promo code (SELLER/ADMIN)
      operationId: createPromoCode
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreateRequest'
      responses:
        '201':
          description: Promo code created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalidError'
        '403':
          $ref: '#/components/responses/AccessDeniedError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ProductIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OrderIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    ProductNotFoundError:
      description: Product not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: PRODUCT_NOT_FOUND
            message: Product not found
    OrderNotFoundError:
      description: Order not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ORDER_NOT_FOUND
            message: Order not found
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: Validation failed
            details:
              price: must be greater than or equal to 0.01
    AccessDeniedError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ACCESS_DENIED
            message: Access denied
    TokenInvalidError:
      description: Token invalid or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid:
              value:
                error_code: TOKEN_INVALID
                message: Token invalid
            expired:
              value:
                error_code: TOKEN_EXPIRED
                message: Token expired
    RefreshTokenInvalidError:
      description: Refresh token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: REFRESH_TOKEN_INVALID
            message: Refresh token invalid

  schemas:
    ErrorCode:
      type: string
      enum:
        - PRODUCT_NOT_FOUND
        - PRODUCT_INACTIVE
        - ORDER_NOT_FOUND
        - ORDER_LIMIT_EXCEEDED
        - ORDER_HAS_ACTIVE
        - INVALID_STATE_TRANSITION
        - INSUFFICIENT_STOCK
        - PROMO_CODE_INVALID
        - PROMO_CODE_MIN_AMOUNT
        - ORDER_OWNERSHIP_VIOLATION
        - VALIDATION_ERROR
        - TOKEN_EXPIRED
        - TOKEN_INVALID
        - REFRESH_TOKEN_INVALID
        - ACCESS_DENIED
        - INTERNAL_ERROR

    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true

    Role:
      type: string
      enum:
        - USER
        - SELLER
        - ADMIN

    ProductStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ARCHIVED

    OrderStatus:
      type: string
      enum:
        - CREATED
        - PAYMENT_PENDING
        - PAID
        - SHIPPED
        - COMPLETED
        - CANCELED

    DiscountType:
      type: string
      enum:
        - PERCENTAGE
        - FIXED_AMOUNT

    UserResponse:
      type: object
      required:
        - id
        - username
        - role
        - created_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time

    AuthRegisterRequest:
      type: object
      required:
        - username
        - password
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        password:
          type: string
          minLength: 6
          maxLength: 100
        role:
          $ref: '#/components/schemas/Role'

    AuthLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        password:
          type: string
          minLength: 6
          maxLength: 100

    AuthRefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          minLength: 16

    AuthTokensResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - access_expires_at
        - refresh_expires_at
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        access_expires_at:
          type: string
          format: date-time
        refresh_expires_at:
          type: string
          format: date-time

    ProductCreate:
      type: object
      required:
        - name
        - price
        - stock
        - category
        - status
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 4000
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductUpdate:
      type: object
      required:
        - name
        - price
        - stock
        - category
        - status
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 4000
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - category
        - status
        - seller_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductPageResponse:
      type: object
      required:
        - content
        - totalElements
        - page
        - size
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
          format: int64
          minimum: 0
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1

    OrderItemRequest:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 999

    OrderCreateRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        promo_code:
          type: string
          nullable: true
          pattern: '^[A-Z0-9_]{4,20}$'

    OrderUpdateRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemResponse:
      type: object
      required:
        - product_id
        - quantity
        - price_at_order
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_order:
          type: number
          format: decimal

    OrderResponse:
      type: object
      required:
        - id
        - user_id
        - status
        - items
        - total_amount
        - discount_amount
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        promo_code:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        total_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PromoCodeCreateRequest:
      type: object
      required:
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - valid_from
        - valid_until
        - active
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
          minimum: 0.01
        min_order_amount:
          type: number
          format: decimal
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean

    PromoCodeResponse:
      type: object
      required:
        - id
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - current_uses
        - valid_from
        - valid_until
        - active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
        min_order_amount:
          type: number
          format: decimal
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time